requirements.txt
Flask==3.0.3
openpyxl==3.1.2
gunicorn==22.0.0
openai>=1.50.0
httpx>=0.27.0
requests>=2.31.0
google-generativeai>=0.8.3
anthropic>=0.39.0
